sql: SHOW CREATE now uses two-part names for enum types
Previously, when viewing DDL through `SHOW CREATE`, enum types were
displayed with a three-part name: `database.schema.type`. This limited
the usefulness of the DDL output because it could only be applied to a
database with the same name.

Now, the `SHOW CREATE` output uses a two-part name (`schema.type`),
matching the style used by postgres in the `pg_dump` command.

There are many test changes due to the numerous uses of the `SHOW
CREATE` command. Most changes involve removing the database name from
the type name.

Additionally, I added an option to `tree/format.go` to allow using a
three-part name for types. This is necessary to continually detect
cross-database type access in views.

Epic: none
Closes: #125827
Release note (bug fix): When showing enum types in `SHOW CREATE` output, we
now format the name as a two-part name instead of a three-part name.
This change makes it easier to apply the output with enum types, to
other databases.