sql/schemachanger: Add support for storing roles in policies
A previous commit introduced basic support for CREATE/DROP POLICY. This
commit extends that functionality by storing roles in the policy
descriptor as strings. If role renaming support is added in the future,
the roles will need to be stored as IDs.

The roles are modelled as separate DSC elements, making it easier to
modify the roles in a policy, which will be supported with ALTER POLICY.

Epic: CRDB-11724
Informs: #136696
Release note: None