sql/schemachanger: Add support for storing expressions in policies
This PR builds on previous work for storing policy info in the table
descriptor. It adds storing of policy expressions. A policy can have
one expression for reading rows (USING) and one expression for writing
rows (WITH CHECK). The necessary plumbing was added to the DSC to allow
us to store the expressions in the table descriptor.

To keep the scope manageable, this PR does not yet handle forward or
backward references for objects (e.g., columns, UDTs, sequences, or
functions) referenced in the expressions. A follow-up PR will address
maintaining these references. Until that is implemented, it will be
possible to perform operations to drop any of these objects when they
are referenced in a policy expression.

Epic: CRDB-11724
Informs: #136696
Release note: None