sql/schemachanger: Add dependency tracking for policy expressions
A prior change introduced the storage of policy expressions within the
table descriptor. This update builds on that by adding forward and
backward references for objects used in these expressions. These
references ensure that dependent objects cannot be dropped if they are
actively used in a policy expression.

As part of this change, functions are now permitted within policy
expressions. Previously, their use was blocked; however, with proper
dependency tracking in place, this restriction can be safely lifted.

To accommodate dependencies arising from the two possible expressions
(USING and WITH CHECK) in a policy, a new descriptor element,
PolicyDeps, was introduced. This element tracks dependencies for both
expressions.

Epic: CRDB-11724
Informs: #136696
Release note: None