sql: Ensure we auto commit for first DDL in txn
Previously, when the autocommit_before_ddl session setting was enabled,
the first DDL statement in a multi-statement transaction did not trigger
an auto-commit, leading to the use of the legacy schema changer. This
behavior was observed in transactions such as:

```
SET autocommit_before_ddl = ON;
BEGIN;
DROP TABLE t1;
DROP TABLE t2;
```

This change will now ensure that we auto commit before the first drop.
This ensures that we'll use the DSC for both drops.

Epic: None
Release note: None

Closes #141034