sql: add metric for transaction timeout failures
Add the sql.transaction_timeout.count metric to track the number of SQL
statements that fail due to exceeding the transaction timeout. This
change is similar to #142078 and serves as a follow-up item from that
work.

Epic: none
Release note (ops change): Adds the sql.transaction_timeout.count metric
to track the number of SQL statements that fail due to exceeding the
transaction timeout.

Informs #142008